v.0.3.

+++ include/matrix_utility.h
@@ -1,9 +1,9 @@
 /**
  * @file matrix_utility.h
- * @author Jonay Estévez Díaz (alu0101100286@ull.edu.es)
+ * @author Jonay Estévez Díaz (alu0101100586@ull.edu.es)
  * @author Bruno Lorenzo Arroyo Pedraza (alu0000000000@ull.edu.es)
- * @author Carla Cristina Olivares Rodriguez (alu000000000@ull.edu.es)
- * @author Jose Miguel Hernández Santana (alu000000000@ull.edu.es)
+ * @author Carla Cristina Olivares Rodriguez (alu0101120218@ull.edu.es)
+ * @author Jose Miguel Hernández Santana (alu0101101507@ull.edu.es)
  * 
  * @brief Clase Container, que nos permite utilizar una matriz como 
  *        container de la matriz de utilidad
@@ -19,8 +19,10 @@
 
 #include <iostream>
 #include <fstream>
+#include <sstream>
 #include <ostream>
 #include <vector>
+#include <math.h>
 
 enum Metrics {Pearson, Cosine, Euclidean};
 enum Prediction {Simple, Mean};
@@ -28,20 +30,20 @@ enum Prediction {Simple, Mean};
 class MatrixUtility {
   public:
     MatrixUtility();
-    MatrixUtility(std::vector<std::vector<int>> users_items);
+    MatrixUtility(std::vector<std::vector<double>> users_items);
     MatrixUtility(std::vector<std::string> users_lines);
     ~MatrixUtility();
 
-    std::vector<std::vector<int>> GetUserItems() const;
+    std::vector<std::vector<double>> GetUserItems() const;
 
-    void SetUserItems(std::vector<std::vector<int>> new_user_items);
+    void SetUserItems(std::vector<std::vector<double>> new_user_items);
 
-    int& At(int i, int j);
+    double& At(int i, int j);
 
     friend std::ostream& operator<<(std::ostream& os, const MatrixUtility m);
 
   private:
-    std::vector<std::vector<int>> user_items_;
+    std::vector<std::vector<double>> user_items_;
     
     std::vector<std::string> Split(std::string line);
 };
